Remove types/winrt-uwp dependency
Added this initially to so that the type definition for UWP APIs
could be downloaded automatically when a project was loaded.
NTVS base project however cannot correctly parse the @types,
instead it tries to do 'npm install winrt-uwp' which is incorrect.
This can be added back later if the issue in NTVS is fixed.